Start working on renderer
We wrote some simple utilities to make code in future simpler. I
created this strange concept "sampler cache". It is supposed to help
us not duplicate samplers. But the more I think the more this cache
looks useless. Also we hardcoded sampler's minLod, maxLod and
borderColor. We would need to extend 'SamplerInfo' struct if we want
to specify this attributes.