Get rid of hash and compare functions for TypeInfo
I realised that generic functions like memhash(which we implemented as
lida_HashMemory32) and memcmp are enough for our purposes. Yes,
hashing could be very bad in some cases, but we doesn't depend on hash
quality. What we really depend on is code simplicity.

Also, I introduced some code for C++ in base.h as part of code base
will be in C++.