Preparing voxel draw optimization
Current draw system works perfectly fine for GPU, but on CPU side it
is a disaster. We generate meshes everyframe for same voxel grids! I
came up with following algorithm:

1. at the start of frame build some accelerating data structures
for later use based on previous frame draws. I was researching which data structure to use. It seems that sorted array is
the way. There're also prefix trees but I don't have experience with them.

2. when 'pushing' a grid to draw fast-lookup if we had this grid with
exact hash in previous frame. If yes then we can reuse generated
vertices.

This algorithm looks good enough for me, there're a bunch of issues
like fragmentation, overwriting previous frame's data, but we'll
handle it.