More new code
I made so many changes that I won't elaborate on them.

However we have a weird bug with hash tables, I should fix it. The worst thing about it that it doesn't crash application.

I'm happy with overall design and hope that I will rewrite engine as possible to work on more features.