Introduce entity component system
We wrote a basic ECS. It's based on sparse sets and should work pretty
fast. However the one thing I dislike about current API is that
iterating over multiple components is slow. Everytime we access
second(or 3rd or 4th or ...) entity's component we call GetSparseSet()
which looks is hash table lookup. Ideally we would achieve the goal
via some prerprocessor magic, but I'm not so good at C macro
trickery. However for the moment this is not an issue; it's a task for
future me :)

Next I will try to add text rendering and then hopefully we can get rid of old code.