Preparations for hot reloading
In each frame we check if any of the files inside 'data' directory are
changed. If yes we process those files.

Currently we only support linux's 'inotify' API. It's a shame that
SDL2 doesn't provide file notification mechanisms. It's still will be
very easy to port to MS Windows.

Now I need to think how we will process those file changes. Initial
thought is to create a hash table which points to entities that
reference files. Need to think more about that.