Add so called 'Keymaps'
In different states a game supposed to accept different
keybindings. For this reason we introduce 'Keymap's.

In Emacs keymap is a sparse map which maps key-sequence strings to
commands. Our keymap are just a set of callbacks which handle
different keymaps. This is much simpler but less configurable as we
don't want to be Emacs.

Also I need to think about how keymaps would inherit keybindings. My
initial though: return 1 if key handled, return 0 if key must be
handled by parent keymaps.