Backface culling💋
This is a really big thing in terms of performance. It basically halfs
number of draw calls we have. Few things we need to be careful with:
- what to do with instancing;
- when moving around I noticed that grid_2's voxel dissapear for a
moment if you look from below. This makes me think that I missed something;

I would also want to implement frustum culling. But we probably won't have time for it.