New voxel caching system🫡
New system doesn't regenerate vertices for voxel models when they
appear. It is also a lot simpler and heavily uses our ECS. I'm
generally happy with this solution but it doesn't handle well some
generic cases when some grids are changed frequently. In fact it will
just run out of memory🫡.

For now I don't care, this is a task for future me.