Improving GPU driven rendering
We're not doing culling yet, but it we made all preparations
needed. This was a bit hard because of my stupid bug. In compute
shader I didn't check if we're out of bounds. At random frames some
mesh information was overwritten and I got horrifying flickering.

Fortunately I solved this🥲