Serialization support for previous commit
It was very easy, but I don't like that each time we introduce some
change in data we need to rewrite serialization code. I'll certainly
use some other language in conjunction in C after the presentation.

Anyway we can duplicate voxel models on screen without duplicating
them in memory which is great 🍰