Trying to fix oclussion culling :(
The main thing that struggles me the most is that I don't make any
progress working on this issue. I decided that I should examine min
and max depths of specific objects in the scene.

A hard but good option seems to implement mouse picking. I'm going to
implement it.